[MOV] product_visible_discount : split a file by models
