+++ addons/product_visible_discount/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
+import product_pricelist
 import sale_order
